separates asset_kitchen into a static library
