+++ include/Lib/asset_kitchen/assets.hpp
@@ -4,7 +4,6 @@
 #include "Lib/opengl/.hpp"
 #include "Lib/render/.hpp"
 #include "Lib/scene/.hpp"
-#include "Lib/asset_kitchen/descriptions.hpp"
 
 #include "descriptions.hpp"
