--- include/Lib/asset_kitchen/glsl/uniform_block/core.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace GLSL::UniformBlock
-{
-	struct LoadedData
-	{
-		std::string source;
-	};
-}
