--- include/Lib/asset_kitchen/glsl/uniform_block/loader.hpp
@@ -1,16 +1,0 @@
-#pragma once
-
-namespace GLSL::UniformBlock
-{
-	struct Description
-	{
-		std::filesystem::path path;
-	};
-
-	LoadedData Load(Description const & description)
-	{
-		return LoadedData{
-			.source = File::LoadAsString(description.path),
-		};
-	}
-}
