+++ include/Lib/scene/core.hpp
@@ -3,6 +3,7 @@
 #include <queue>
 
 #include "Lib/core/core.hpp"
+#include "Lib/core/utils.hpp"
 #include "Lib/render/mesh.hpp"
 
 namespace Scene
