+++ source/main.cpp
@@ -7,6 +7,14 @@
 #include "game.hpp"
 #include "editor.hpp"
 
+// TODO(bekorn): move this back into opengl folder
+// To run the application with the external GPU, https://stackoverflow.com/a/39047129/2073225
+extern "C"
+{
+__declspec(dllexport) i64 NvOptimusEnablement = 1;
+__declspec(dllexport) i32 AmdPowerXpressRequestHighPerformance = 1;
+}
+
 i32 main(i32 argc, char** argv)
 {
 	if (argc < 2)
